Gitlab-CI: Fix running tests
If the plugin is not installed then tshark ignores the -Xwslua2
option extension and the test suite fails silently. Make sure
the plugin is installed before running the test suite.

There is also a bug in Wireshark preventing "tshark --version"
from displaying the number of plugins loaded that will be fixed
upstream.